Updated data pipeline
