--- utils/object.py
@@ -1,8 +1,0 @@
-class MLModelConfig:
-    def __init__(self, X, y, X_train, X_test, y_train, y_test):
-        self.X = X
-        self.y = y
-        self.X_train = X_train
-        self.X_test = X_test
-        self.y_train = y_train
-        self.y_test = y_test
